feat(F0Chat): nest the reply quote inside the message bubble
Move the reply quote from a separate card above the bubble to a nested
card at the top of the bubble (above the body), WhatsApp-style. A quoted
message now always renders a bubble to host it.
